ASoC:[Backported msm-4.4r9] wcd-dsp-mgr: add support for subsystem restart
It is possible that wcd codec DSP can crash or become unresponsive.
During such case, an error interrupt is generated by the codec. Add
support in manager driver to handle this interrupt and perform
subsystem restart to shutdown and reboot the DSP so that the DSP
can be recovered from the crash.

CRs-fixed: 1071949
Change-Id: I4662b5120bf7f731e399a27d8a613e2f3b648b00
Signed-off-by: Bhalchandra Gajare <gajare@codeaurora.org>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>
Signed-off-by: Dmitrii <bankersenator@gmail.com>