ASoC:[Backported msm-4.4r9] wcd-dsp-mgr: fix restart logic when codec comes back online
During subsystem restart, when restart of the codec DSP is attempted,
the status is already moved to codec up whereas driver is checking
for codec down. This causes the initialization of components to be
skipped and codec dsp does not come back online. Change fixes the
condition check to make sure the correct sequence is followed.

CRs-Fixed: 1080392
Change-Id: I05345b8f5d108e59863ee4e37ed75fd32a284ee5
Signed-off-by: Bhalchandra Gajare <gajare@codeaurora.org>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>
Signed-off-by: Dmitrii <bankersenator@gmail.com>